Minor update
- Turtlecoin v0.18.0 Latest release

+++ src/zedwallet++/CommandImplementations.cpp
@@ -405,7 +405,7 @@ void printIncomingTransfer(const WalletTypes::Transaction tx)
                   << std::endl;
     }
     /* Here we treat Unlock as Unix time, and treat it that way in the future */
-    else if (tx.unlockTime > std::time(nullptr))
+    else if (tx.unlockTime > static_cast<uint64_t>(std::time(nullptr)))
     {
         std::cout << SuccessMsg(stream.str()) << InformationMsg("Unlocks at: ")
                   << InformationMsg(Utilities::unixTimeToDate(tx.unlockTime)) << std::endl
